Add caching:
We now have a context struct (MY_CXT) which holds to a hash we
use in order to cache results to the function.

This could be done with basic C structures instead of Perl
structures (which would make it much faster) but it's easy to at
least do it this way for now.