Normalize:
Reproducing %standard_case would take a while (and I'm still not
comfortable with doing it in C with Perl structures) so for now
we can just make it a global we can now reach.

The logic is as follows:

If a header is not in the %standard_case hash, rewrite it to have
an uppercase letter before each word boundary. Then store it back
in the hash.

This transforms "x-foo-bar" to "X-Foo-Bar".

The original code in Perl:

    $old =~ s/\b(\w)/\u$1/g;